拆分Point
