+++ src/manimgeo/components/__init__.py
@@ -1,5 +1,5 @@
 from ..components.base import *
-from .point.point import *
+from .point import Point
 from .line.line import *
 from .vector.vector import *
 from .circle.circle import *
