+++ src/manimgeo/components/point/__init__.py
@@ -1,0 +1,2 @@
+from .point import Point
+from .adapter import PointAdapter
